refactor: create fixture for function making different vault versions
also implement caching on it to speed it up
